fix: 엔티티에 맞게 summary 수정

+++ src/main/java/com/team3/monew/document/CommentLikeSummary.java
@@ -12,7 +12,7 @@ public record CommentLikeSummary(
     UUID commentUserId,
     String commentUserNickname,
     String commentContent,
-    Long commentLikeCount,
+    Integer commentLikeCount,
     Instant commentCreatedAt
 ) {
